pyproject.toml
[project]
name = "genie-toolkit"
version = "0.1.0"
description = "Genie toolkit to help with genie based projects."
readme = "README.md"
authors = [
]
requires-python = ">=3.11"
dependencies = [
    "databricks-sdk>=0.77.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "typer>=0.21.1",
    "gepa>=0.1.0",
]

[project.scripts]
genie = "genie_toolkit.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.11",
    "pytest-cov>=7.0.0",
]

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"
